fix: address all CodeQL High/Error findings
Multiplication overflow (#172)
  friendlistmanager_test.cpp: cast dayRatio to qint64 before int*int
  multiplication to prevent overflow before widening.

File created without restricting permissions (#150, #151)
  appmanager.cpp: replace both fopen("a") log-file calls with a
  POSIX open(O_CREAT|O_APPEND, 0600)+fdopen() lambda on non-Windows,
  ensuring the log file is created with user-only permissions.
  Add #include <fcntl.h> under #ifndef Q_OS_WIN.

Incorrect allocation-error handling (#149)
  openal.cpp: change 'new AlSink' to 'new (std::nothrow) AlSink' so
  the existing nullptr guard is meaningful rather than dead code.

Missing return statement (#103, #104)
  debugobjecttreemodel.cpp, rawdatabaseimpl.h: add Q_UNREACHABLE()
  after qFatal() so the compiler and analysers know the path is
  unreachable and no return value is required.

Array argument size mismatch (#148)
  settingsserializer.cpp: replace tox_is_data_encrypted() with an
  inline memcmp against the 8-byte "toxEsave" magic; remove the now-
  unused <tox/toxencryptsave.h> include. Also use sizeof(magic) instead
  of the literal 4 for the qTox magic comparison.

Inconsistent nullness check (#174-#180)
  conferenceroom.cpp:  guard inviteFriend() against null findFriend()
  friendchatroom.cpp:  guard removeFriendFromDialogs() against null getFriendDialogs()
  contentdialog.cpp:   guard updateFriendWidget() against null findFriend()
  contentdialogmanager.cpp: guard updateFriendStatus() against null findFriend()
  widget.cpp:          guard newFriendMessageAlert() and
                       onConferenceInviteReceived() against null findFriend()
  friendlistwidget.cpp: make null check explicit before delete in dtor

Virtual call from constructor/destructor (#105, #106)
  genericchatform.cpp: call GenericChatForm::reloadTheme() explicitly
    to suppress virtual dispatch from the base-class constructor.
  categorywidget.cpp:  inline the setExpanded(true,false) body directly
    in the constructor to eliminate the indirect path through the virtual
    onExpand(); update initial statusPic to the expanded (down-arrow) pixmap.

+++ test/model/friendlistmanager_test.cpp
@@ -336,7 +336,7 @@ private:
             return dateTime;
         }
         const int dayRatio = -1;
-        return dateTime.addDays(dayRatio * pos * pos);
+        return dateTime.addDays(static_cast<qint64>(dayRatio) * pos * pos);
     }
 
     QStringList sortedByNameOfflineFriends;
